Disable evictions pathway (#18)
* Resolve "error:0308010C:digital envelope routines::unsupported"

Node v17.0.0 changed to use OpenSSL 3.0 which no longer allows certain hashing
algorithms by default. This breaks the version of webpack and, as a
consequence, storybook we are currently using as it uses one of the 
now-disallowed algorithms (MD4). This is fixed in later versions of webpack where
support was added to generate MD4 hashs directly rather than via OpenSSL.

To resolve this in webpack I have updated to the earliest version that contains
their fix, v4.47.0. This version was chosen rather than a more recent version
so we don't have to update all of our dependencies at this time.

I have used a different approach for storybook as it would require considerable
time and effort to update to a version that uses a fixed version of webpack. We
just allow OpenSSL to use "legacy" algorithms using the
NODE_OPTIONS=--openssl-legacy-provider environment variable.

See for discussion:

- https://stackoverflow.com/a/73027407
- https://github.com/webpack/webpack/issues/14532
- https://nodejs.org/en/blog/release/v17.0.0#openssl-30

See for fix in webpack:

- https://github.com/webpack/webpack/pull/14584
- https://github.com/webpack/webpack/releases/tag/v4.47.0