Fix react-snap hang
The ultimate error that causes the hang seems to be coming from chrome via
puppeteer:

        FATAL:gpu_data_manager_impl_private.cc(918)] The display compositor is
        frequently crashing. Goodbye.

This might be resolved in chrome that ships with later puppeteer versions but
react-snap is already at the latest version. Given the limited use of
react-snap on trusted web content, disabling the chrome sandbox seems
reasonable to get this working.