Kludge fix for buttons being clipped on some systems
When the office closed notice is displayed the buttons to start form submission
are hidden on OSX but not Linux (not tested on Windows). The buttons are hidden
as they are outside of the viewport height.

It looks like this is due to the text being larger on OSX than Linux, possibly
due to the available fonts? This change prevents content clipping which allows
us to scroll to see content outside of the viewport height.