Prevent error fetching clerk API when also logged into clerk
The error is as follows:

  CSRF Failed: Origin checking failed - http://localhost:3001 does not match any trusted origins.

The issue results because we are including credentials in the fetch request
which, when logged in to clerk, include a "sessionid" and "csrftoken" cookie
values. Including a sessionid means that we would also have to include the
valid csrftoken but this is unnecessary as we are hitting an unauthenticated
API endpoint anyway.

We just revert to use the default fetch credentials setting of same-origin;
only send and include credentials for same-origin requests. See
https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch

Also we remove the confusing empty CSRF token header. Presumably this was valid
at some point but it is no longer necessary.