Restrict access to necessary S3 buckets only
The previous access key pair allows full access to all of our S3 assets. It is
more secure to use finer-grained access controls in case the keys are
compromised.

In this case the access keys are only used as part of the intake form
deployment so it's not particularly problematic but I want to retire the key
pair that allows full access so we have to change the keys.

This change required that we set up AWS users with permission policies that
only allow access to specific S3 buckets, assign them access keys and use those
access keys here.